feat: use know fee component as stylized and reusable fee field
