+++ src/ui/view/share/knowFee/knowFee.scss
@@ -13,21 +13,22 @@
     border-radius: 12px;
     height: 128px;
     color: themed('text');
-
+    
     &.dark {
       @include background-image('dark');
     }
-
+    
     &.light {
       @include background-image('light');
     }
-
-    p {
+    
+    .okp4-dataverse-portal-share-data-know-fee-label {
+      display: block;
       @include noto-sans(400);
       box-sizing: border-box;
       padding: 18px 21px;
     }
-
+    
     .okp4-dataverse-portal-share-data-know-fee-field {
       width: 250px;
       padding-left: 21px;
